+++ src/main/resources/templates/review/review-update.html
@@ -104,7 +104,7 @@
 <script type="text/javascript" src="//dapi.kakao.com/v2/maps/sdk.js?appkey=59f5c92e8da50255a30ef20d44e71dcf&libraries=services"></script>
 <script src="https://cdnjs.cloudflare.com/ajax/libs/lodash.js/4.17.21/lodash.min.js" integrity="sha512-WFN04846sdKMIP5LKNphMaWzU7YpMyCU245etK3g/2ARYbPK9Ub18eG+ljU96qKRCWh+quCY7yefSmlkQw1ANQ==" crossorigin="anonymous" referrerpolicy="no-referrer"></script>
 
-<script src="/assets/js/review-update.js"></script>
+<script type="module" src="/assets/js/review-update.js"></script>
 
 </body>
 </html>
