+++ src/main/resources/templates/review/review-write.html
@@ -137,7 +137,7 @@
 <!-- kakao maps api -->
 <script type="text/javascript" src="//dapi.kakao.com/v2/maps/sdk.js?appkey=59f5c92e8da50255a30ef20d44e71dcf&libraries=services"></script>
 
-<script src="/assets/js/review-write.js"></script>
+<script type="module" src="/assets/js/review-write.js"></script>
 
 </body>
 </html>
